pyproject.toml
[tool.poetry]
name = "fastapi-course"
version = "0.1.0"
description = ""
authors = ["Дарья Коннова <dashustic@yandex.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = {extras = ["all"], version = "^0.104.1"}
uvicorn = {extras = ["standart"], version = "^0.24.0.post1"}
sqlalchemy = {extras = ["asyncio"], version = "^2.0.23"}
pydantic = {extras = ["email"], version = "^2.5.2"}
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
pyjwt = {extras = ["crypto"], version = "^2.8.0"}
asyncpg = "^0.29.0"
alembic = "^1.13.1"


[tool.poetry.group.dev.dependencies]
black = "^23.11.0"
pytest = "^7.4.3"
isort = "^5.13.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
